plain-observer/pyproject.toml
[project]
name = "plain.observer"
version = "0.1.0"
description = "Basic observability and monitoring tools for Plain."
authors = [{name = "Dave Gaeddert", email = "dave.gaeddert@dropseed.dev"}]
license = "BSD-3-Clause"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "opentelemetry-sdk>=1.34.1",
    "plain<1.0.0",
    "plain.auth<1.0.0",
    "plain.tailwind<1.0.0",
    "sqlparse>=0.2.2",
]

[tool.uv]
dev-dependencies = [
    "plain.pytest<1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["plain"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
